chore: Update registration of PostInstall workflow in mothership main.go
This commit updates the registration of the PostInstall workflow in the mothership main.go file. The workflow is now registered as RefreshDefaultBranch to accurately reflect its purpose of refreshing the default branch for all repositories associated with the given teamID. This change ensures that the workflow is properly registered and can be triggered when needed.